fix: Corrigindo a criação do veículo e associando o acessório pelo frontend
